refactor: вынесение хранения данных о фильме в отдельный класс
